backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
python-dotenv
